太っていたapp.dartをwidgetとdomainに分割
