+++ flutter/lib/app/app.dart
@@ -1,0 +1,24 @@
+import 'package:flutter/material.dart';
+import 'package:firebase_analytics/firebase_analytics.dart';
+import 'package:firebase_analytics/observer.dart';
+
+import 'package:bacchus/app/pages/home.dart';
+
+class MyApp extends StatelessWidget {
+  static FirebaseAnalytics analytics = FirebaseAnalytics();
+
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      title: 'Bacchus',
+      theme: ThemeData(
+        primarySwatch: Colors.blue,
+      ),
+      home: HomePage(title: 'Home'),
+      navigatorObservers: [
+        FirebaseAnalyticsObserver(analytics: analytics),
+      ],
+    );
+  }
+}
+
