+++ flutter/lib/main.dart
@@ -1,4 +1,4 @@
-import 'ui/app.dart';
+import 'app/app.dart';
 import 'package:flutter/material.dart';
 
 void main() => runApp(MyApp());
